Fix integer overflow in convertVectorToInt.

+++ libzerocash/utils/util.cpp
@@ -162,7 +162,7 @@ uint64_t convertVectorToInt(const std::vector<bool>& v) {
     uint64_t result = 0;
     for (size_t i=0; i<v.size();i++) {
         if (v.at(i)) {
-            result |= 1 << ((v.size() - 1) - i);
+            result |= (uint64_t)1 << ((v.size() - 1) - i);
         }
     }
